+++ CHANGELOG.md
@@ -14,7 +14,7 @@ These changes are available on the `master` branch, but have not yet been releas
 
 - Added `Member.vr_status` property.
   ([#3328](https://github.com/Pycord-Development/pycord/pull/3328))
-- Added spoiler channel functionality
+- Added spoiler channels.
   ([#3252](https://github.com/Pycord-Development/pycord/pull/3252))
 
 ### Changed